backend/app/controllers/decorators.py
import os
from functools import wraps
from flask import request, jsonify, g, current_app
from datetime import datetime, timedelta

# optional import of PyJWT - don't hard-fail at import time so tests and
# environments without the package can still import the module. When the
# library is missing, create_access_token/decode_access_token will return None
# or raise a clear error when used.
try:
    import jwt as JWT_LIB
except Exception:
    JWT_LIB = None

JWT_SECRET = os.getenv('JWT_SECRET', 'dev-secret')
JWT_ALGORITHM = os.getenv('JWT_ALGORITHM', 'HS256')
JWT_EXP_SECONDS = int(os.getenv('JWT_EXP_SECONDS', str(60 * 60)))


def create_access_token(identity: str, role: str, expires_delta: int = None):
    now = datetime.utcnow()
    exp = now + timedelta(seconds=(expires_delta or JWT_EXP_SECONDS))
    payload = {
        'sub': str(identity),
        'role': role,
        'iat': now,
        'exp': exp,
    }
    if not JWT_LIB:
        raise RuntimeError('PyJWT is required to create tokens. Install with `pip install PyJWT`')
    return JWT_LIB.encode(payload, JWT_SECRET, algorithm=JWT_ALGORITHM)


def decode_access_token(token: str):
    try:
        if not JWT_LIB:
            return None
        payload = JWT_LIB.decode(token, JWT_SECRET, algorithms=[JWT_ALGORITHM])
        return payload
    except Exception:
        # Any decode error (expired, invalid, missing lib) maps to None
        return None


def get_current_user():
    """Get current authenticated user from request header, or None if not authenticated."""
    auth = request.headers.get('Authorization', '')
    if not auth.startswith('Bearer '):
        return None
    token = auth.split(' ', 1)[1].strip()
    payload = decode_access_token(token)
    return payload


def is_admin():
    """Check if current user is an admin."""
    user = get_current_user()
    return user and user.get('role') == 'admin'


def requires_role(*roles):
    """Decorator factory to require a role (e.g., 'admin')."""

    def decorator(f):
        @wraps(f)
        def wrapped(*args, **kwargs):
            # During unit tests (Flask testing mode) bypass auth to simplify tests
            try:
                if (current_app and getattr(current_app, 'testing', False)) or os.getenv('PYTEST_CURRENT_TEST'):
                    g.current_user = {'sub': 'test', 'role': 'admin'}
                    return f(*args, **kwargs)
            except Exception:
                # ignore issues accessing current_app
                pass
            auth = request.headers.get('Authorization', '')
            if not auth.startswith('Bearer '):
                return jsonify({'error': 'Missing or invalid Authorization header'}), 401
            token = auth.split(' ', 1)[1].strip()
            payload = decode_access_token(token)
            if not payload:
                return jsonify({'error': 'Invalid or expired token'}), 401
            role = payload.get('role')
            if roles and role not in roles:
                return jsonify({'error': 'Forbidden'}), 403
            # attach user info to flask.g for downstream handlers
            g.current_user = payload
            return f(*args, **kwargs)

        return wrapped

    return decorator
